rename all error_codes objects to ec
Keeping error_codes with a consistant name makes the code easier to
read. While "status" and "ret" was a good substitution for the
somethimes context-confusing "error", it didn't hold the same
self-documenting as:

    if (error)
        // do thing with error

"ec" is used all over Boost's examples, and also holds the
self-documentation of "error", so we're moving to that.